requirements.txt
asyncio            
base58
indy-vdr
pycparser
PyNaCl
python-dotenv
requests